RF: bring common (?) functionality under PathBasedFlyweight
We had to almost identical copies of the code for _flyweight_id_from_args .
That can breed bugs having fixes in one copy but not another.  I have decided
to go superclass subclassing path so we could also "modularize" the logic for
handling aliases (ATM only for the datasets).

I have taken implementation from  Dataset, which did not explicitly
"realpath" the path.  Not sure yet if such difference in behavior (Dataset vs *Repo)
was intended/desired, tests might show
