+++ datalad/distribution/dataset.py
@@ -36,7 +36,7 @@ from datalad.support.external_versions import external_versions
 from datalad.support.gitrepo import GitRepo
 from datalad.support.gitrepo import InvalidGitRepositoryError
 from datalad.support.gitrepo import NoSuchPathError
-from datalad.support.repo import Flyweight
+from datalad.support.repo import PathBasedFlyweight
 from datalad.support.network import RI
 from datalad.support.exceptions import InvalidAnnexRepositoryError
 
@@ -79,7 +79,7 @@ def resolve_path(path, ds=None):
     return normpath(opj(top_path, path))
 
 
-@add_metaclass(Flyweight)
+@add_metaclass(PathBasedFlyweight)
 class Dataset(object):
     """Representation of a DataLad dataset/repository
 
@@ -104,26 +104,8 @@ class Dataset(object):
     _unique_instances = WeakValueDictionary()
 
     @classmethod
-    def _flyweight_id_from_args(cls, *args, **kwargs):
-
-        if args:
-            # to a certain degree we need to simulate an actual call to __init__
-            # and make sure, passed arguments are fitting:
-            # TODO: Figure out, whether there is a cleaner way to do this in a
-            # generic fashion
-            assert('path' not in kwargs)
-            path = args[0]
-            args = args[1:]
-        elif 'path' in kwargs:
-            path = kwargs.pop('path')
-        else:
-            raise TypeError("__init__() requires argument `path`")
-
-        if path is None:
-            lgr.debug("path is None. args: %s, kwargs: %s", args, kwargs)
-            raise ValueError("path must not be None")
-
-        # Custom handling for few special abbreviations
+    def _flyweight_preproc_path(cls, path):
+        """Custom handling for few special abbreviations for datasets"""
         path_ = path
         if path == '^':
             # get the topmost dataset from current location. Note that 'zsh'
@@ -135,20 +117,7 @@ class Dataset(object):
             path_ = cfg.obtain('datalad.locations.default-dataset')
         if path != path_:
             lgr.debug("Resolved dataset alias %r to path %r", path, path_)
-
-        # Sanity check for argument `path`:
-        # raise if we cannot deal with `path` at all or
-        # if it is not a local thing:
-        path_ = RI(path_).localpath
-
-        # we want an absolute path, but no resolved symlinks
-        if not isabs(path_):
-            path_ = opj(getpwd(), path_)
-
-        # use canonical paths only:
-        path_ = normpath(path_)
-        kwargs['path'] = path_
-        return path_, args, kwargs
+        return path_
     # End Flyweight
 
     def __init__(self, path):
